allow build with local taserver.zip

+++ build/install_taserver.sh
@@ -3,12 +3,16 @@
 # This script installs Tribes and taserver during docker build
 set -ex
 
-TASERVER_RELEASE_TAG="0.0.32" 
+TASERVER_RELEASE_TAG="0.0.35"
 
 # get taserver
-wget -q -O taserver.zip "https://github.com/chickenbellyfin/taserver/archive/refs/tags/$TASERVER_RELEASE_TAG.zip"
-unzip -q taserver.zip
-mv "taserver-$TASERVER_RELEASE_TAG" "taserver"
+if [ ! -f taserver.zip ]; then
+  wget -q -O taserver.zip "https://github.com/chickenbellyfin/taserver/archive/refs/tags/$TASERVER_RELEASE_TAG.zip"
+  unzip -q taserver.zip
+  mv "taserver-$TASERVER_RELEASE_TAG" "taserver"
+else
+  unzip -q taserver.zip
+fi
 
 cd taserver
